Only install powerline when not installed yet

+++ powerline/install.sh
@@ -10,15 +10,24 @@
 ######################################################
 
 
-su -c 'pip install git+git://github.com/Lokaltog/powerline'
+sudo pip install git+git://github.com/Lokaltog/powerline
 
 TMP_DIR=`mktemp -d`
 pushd $TMP_DIR
 
-wget https://github.com/Lokaltog/powerline/raw/develop/font/PowerlineSymbols.otf https://github.com/Lokaltog/powerline/raw/develop/font/10-powerline-symbols.conf
-sudo mv PowerlineSymbols.otf /usr/share/fonts/
-sudo fc-cache -vf
-sudo mv 10-powerline-symbols.conf /etc/fonts/conf.d/
+if [ ! -f /usr/share/fonts/PowerlineSymbols.otf ]
+then
+    wget https://github.com/Lokaltog/powerline/raw/develop/font/PowerlineSymbols.otf https://github.com/Lokaltog/powerline/raw/develop/font/10-powerline-symbols.conf
+    sudo mv PowerlineSymbols.otf /usr/share/fonts/
+    sudo fc-cache -vf
+fi
+
+if [ ! -f /etc/fonts/conf.d/10-powerline-symbols.conf ]
+then
+    sudo mv 10-powerline-symbols.conf /etc/fonts/conf.d/
+fi
 
 popd
+
+rm -rf $TMP_DIR
